PHP interfaces moved, compare moved, full list of comparison available on all Shoop types
