+++ src/Interfaces/ShoopedExtensions/PhpMagicMethods.php
@@ -1,0 +1,14 @@
+<?php
+
+namespace Eightfold\Shoop\Interfaces\ShoopedExtensions;
+
+interface PhpMagicMethods
+{
+    public function __call(string $name, array $args = []);
+
+    public function __get($name);
+
+    public function __toString(): string;
+
+    public function __debugInfo();
+}
