+++ src/Traits/CompareImp.php
@@ -13,31 +13,31 @@ use Eightfold\Shoop\{
 
 trait CompareImp
 {
-    public function isGreaterThan($compare, \Closure $closure = null)
-    {
-        $compare = Type::sanitizeType($compare, static::class);
-        $bool = $this->unfold() > $compare->unfold();
-        return $this->condition($bool, $closure);
-    }
+    // public function isGreaterThan($compare, \Closure $closure = null)
+    // {
+    //     $compare = Type::sanitizeType($compare, static::class);
+    //     $bool = $this->unfold() > $compare->unfold();
+    //     return $this->condition($bool, $closure);
+    // }
 
-    public function isGreaterThanOrEqual($compare, \Closure $closure = null)
-    {
-        $compare = Type::sanitizeType($compare, static::class);
-        $bool = $this->unfold() >= $compare->unfold();
-        return $this->condition($bool, $closure);
-    }
+    // public function isGreaterThanOrEqual($compare, \Closure $closure = null)
+    // {
+    //     $compare = Type::sanitizeType($compare, static::class);
+    //     $bool = $this->unfold() >= $compare->unfold();
+    //     return $this->condition($bool, $closure);
+    // }
 
-    public function isLessThan($compare, \Closure $closure = null)
-    {
-        $compare = Type::sanitizeType($compare, static::class);
-        $bool = $this->unfold() < $compare->unfold();
-        return $this->condition($bool, $closure);
-    }
+    // public function isLessThan($compare, \Closure $closure = null)
+    // {
+    //     $compare = Type::sanitizeType($compare, static::class);
+    //     $bool = $this->unfold() < $compare->unfold();
+    //     return $this->condition($bool, $closure);
+    // }
 
-    public function isLessThanOrEqual($compare, \Closure $closure = null)
-    {
-        $compare = Type::sanitizeType($compare, static::class);
-        $bool = $this->unfold() <= $compare->unfold();
-        return $this->condition($bool, $closure);
-    }
+    // public function isLessThanOrEqual($compare, \Closure $closure = null)
+    // {
+    //     $compare = Type::sanitizeType($compare, static::class);
+    //     $bool = $this->unfold() <= $compare->unfold();
+    //     return $this->condition($bool, $closure);
+    // }
 }
