+++ tests/Shooped/IsLessThanTest.php
